// frontend/src/types/backend.d.ts
export interface IBackendRes<T> {
    error?: string | string[]
    message: string
    statusCode: number | string
    data?: T
}
export interface IDataWithCodeId {
    codeId: string // Mã codeId trong phản hồi
}

export interface IModelPaginate<T> {
    meta: {
        current: number
        pageSize: number
        pages: number
        total: number
    }
    result: T[]
}

export interface IAccount {
    access_token: string
    user: {
        company: null
        _id: string
        email: string
        name: string
        age: string
        gender: string
        address: string
        role: {
            _id: string
            name: string
        }
        permissions: {
            _id: string
            name: string
            apiPath: string
            method: string
            module: string
        }[]
    }
}

export interface IGetAccount extends Omit<IAccount, "access_token"> {}

export interface ICompany {
    _id?: string
    name?: string
    address?: string
    logo: string
    description?: string
    createdBy?: string
    isDeleted?: boolean
    deletedAt?: boolean | null
    createdAt?: string
    updatedAt?: string
}

export interface IUser {
    _id?: string
    name: string
    email: string
    password?: string
    age: number
    gender: string
    address: string
    role?: {
        _id: string
        name: string
    }

    company?: {
        _id: string
        name: string
    }
    createdBy?: string
    isDeleted?: boolean
    deletedAt?: boolean | null
    createdAt?: string
    updatedAt?: string
}

export interface IJob {
    _id?: string
    name: string
    skills: string[]
    company?: {
        _id: string
        name: string
        logo?: string
    }
    location: string
    salary: number
    quantity: number
    level: string
    description: string
    startDate: Date
    endDate: Date
    isActive: boolean

    createdBy?: string
    isDeleted?: boolean
    deletedAt?: boolean | null
    createdAt?: string
    updatedAt?: string
}

export interface IResume {
    _id?: string
    email: string
    userId: string
    url: string
    status: string
    companyId:
        | string
        | {
              _id: string
              name: string
              logo: string
          }
    jobId:
        | string
        | {
              _id: string
              name: string
          }
    history?: {
        status: string
        updatedAt: Date
        updatedBy: { _id: string; email: string }
    }[]
    createdBy?: string
    isDeleted?: boolean
    deletedAt?: boolean | null
    createdAt?: string
    updatedAt?: string
}

export interface IPermission {
    _id?: string
    name?: string
    apiPath?: string
    method?: string
    module?: string

    createdBy?: string
    isDeleted?: boolean
    deletedAt?: boolean | null
    createdAt?: string
    updatedAt?: string
}

export interface IRole {
    _id?: string
    name: string
    description: string
    isActive: boolean
    permissions: IPermission[] | string[]

    createdBy?: string
    isDeleted?: boolean
    deletedAt?: boolean | null
    createdAt?: string
    updatedAt?: string
}

export interface ISubscribers {
    _id?: string
    name?: string
    email?: string
    skills: string[]
    createdBy?: string
    isDeleted?: boolean
    deletedAt?: boolean | null
    createdAt?: string
    updatedAt?: string
}

export interface IProvince {
    code: string
    name: string
}
export interface INotification {
    _id?: string
    userId: string
    message: string
    isRead: boolean
    createdAt: Date
}